*Fixed some issues *Added 16:9 forwarder *Changed video mode change behaviour *Added new GameSetting "Online fix"
NOTE: Online fix is ONLY needed if you initialize network inside the loader and start an Online Game and ONLY if your boot cios is the same as GameSettings IOS (which is default). That Game would need an IOS Reload before start to be able to play it online BUT this IOS Reload mostly causes Blackscreens. I made this option so you dont have to avoid initializing network. If you have different boot IOS than gamesettings IOS this fix doesnt matter too.

git-svn-id: http://usbloader-gui.googlecode.com/svn/trunk@511 634f2604-25eb-11de-a8df-a7b4aff12ff9