*Added Alternate DOL support which is used to boot games that don't work due to IOS Reloads like Mortal Kombat
NOTE: There is a GameSettings you need to turn ON to load Alternate DOLs. The alternate DOL has to be called 6 IDs .dol for example RHDP8P.dol and has to be in the root of the SD. Giantpune will add custom path support i dont have time for it.

git-svn-id: http://usbloader-gui.googlecode.com/svn/trunk@528 634f2604-25eb-11de-a8df-a7b4aff12ff9