put in a possible "fix"  for the disappearing cursors that some people are getting.  when menu_disclist is called the cursors are deleted and recreated.  so at least if they do go away, this may make them come back.
make the homebrew menu read the XML data from meta.xml file.  put this data on the homebrew buttons and prompt.  there is 4 structs created globally that hold the XML daa.  i tried to make them pointers but i got errors so i'll leave it like this and let dimok fix it :P

git-svn-id: http://usbloader-gui.googlecode.com/svn/trunk@633 634f2604-25eb-11de-a8df-a7b4aff12ff9