*code beautification* formatted the code to make it easier to read.  no functional changes at all.
i didn't put anything from the libwiigui folder or banner folder in the beautifier.

my automated .bat seems to have done a good job.  the only places i see it fucked up was on (GXColor){blablabla}.  it treated the brackets in the color like all the other brackets and put the color on a new line and indented it.  i think i fixed most of them.  not sure if it messed up anywhere else.  also not sure about how it handled different linebreaks.  it looks fine on windows.  if it looks messed up on linux, it can be reverted.

the code still compiles and runs fine.

git-svn-id: http://usbloader-gui.googlecode.com/svn/trunk@675 634f2604-25eb-11de-a8df-a7b4aff12ff9