Made game ID text into a button to change the ID for custom games based off a retail disc (Guitar Hero baby!).
*NOTE*
it is possible to change any game ID to more or less than 6 characters, but since most of the functions use header->id[0] through [5], more will be useless and less may result in code dumps.

it is not advised to change the ID for games that are not custom, as the cover and wiitdb data will not be available for made up ID

it is possible to have multiple games with the same ID on the wbfs partition, but when any of them are clicked, the wii will perform the selected action (play, delete, etc) on the first one it finds.

git-svn-id: http://usbloader-gui.googlecode.com/svn/trunk@719 634f2604-25eb-11de-a8df-a7b4aff12ff9